amelioration admin album

+++ classes/modele/modele_dao/AlbumDAO.class.php
@@ -945,21 +945,25 @@ class AlbumDAO extends DAO {
 		return $this->rrmdir($dir_nom);
 	}
 
-	private function rrmdir($dir) { 
-		if (is_dir($dir)){
-			$objects = scandir($dir);
-			foreach ($objects as $object){
-				if ($object != "." && $object != ".."){
-					if (filetype($dir.DIRECTORY_SEPARATOR.$object) == "dir"){
-						$this->rrmdir($dir.DIRECTORY_SEPARATOR.$object);
-					}else{
-						unlink($dir.DIRECTORY_SEPARATOR.$object);
+	private function rrmdir($dir) {
+		if(file_exists($dir)){
+			if (is_dir($dir)){
+				$objects = scandir($dir);
+				foreach ($objects as $object){
+					if ($object != "." && $object != ".."){
+						if (filetype($dir.DIRECTORY_SEPARATOR.$object) == "dir"){
+							$this->rrmdir($dir.DIRECTORY_SEPARATOR.$object);
+						}else{
+							unlink($dir.DIRECTORY_SEPARATOR.$object);
+						}
 					}
 				}
 			}
+			reset($objects);
+			return rmdir($dir);
+		}else{
+			return true;
 		}
-		reset($objects);
-		return rmdir($dir);
 	}
 
 	private function listeAlbumValidee($liste){
